do not show anything

+++ app/src/main/java/com/indiewalk/mystic/weatherapp/utilities/WeatherAppGenericUtility.java
@@ -97,7 +97,7 @@ public final class WeatherAppGenericUtility {
      * @return Wind String in the following form: "2 km/h SW"
      * ---------------------------------------------------------------------------------------------
      */
-    public static String getFormattedWind(Context context, float windSpeed, float degrees) {
+    public static String getFormattedWind(Context context, double windSpeed, double degrees) {
 
         int windFormat = R.string.format_wind_kmh;
